+++ ZaloPay/app/src/main/java/vn/com/vng/zalopay/react/ReactTransactionLogsNativeModule.java
@@ -215,10 +215,10 @@ class ReactTransactionLogsNativeModule extends ReactContextBaseJavaModule implem
         item.putDouble("transid", history.transid);
         item.putDouble("reqdate", history.reqdate);
         item.putString("description", history.description);
-        item.putInt("amount", history.amount);
-        item.putInt("userfeeamt", history.userfeeamt);
-        item.putInt("type", history.type);
-        item.putInt("sign", history.sign);
+        item.putDouble("amount", history.amount);
+        item.putDouble("userfeeamt", history.userfeeamt);
+        item.putDouble("type", history.type);
+        item.putDouble("sign", history.sign);
         item.putString("username", history.username);
         item.putString("appusername", history.appusername);
         return item;
