+++ ZaloPay/data/src/main/java/vn/com/vng/zalopay/data/api/response/redpacket/SentBundleListResponse.java
@@ -16,7 +16,7 @@ public class SentBundleListResponse extends BaseResponse {
     public long totalOfSentAmount;
 
     @SerializedName("totalofsentbundle")
-    public int totalOfSentBundle;
+    public long totalOfSentBundle;
 
     @SerializedName("sentbundlelist")
     public List<SentBundleResponse> bundleResponseList;
